TECH-869: added count to the pipeline
